Move CollectionChangeBuilder to background_collection.hpp
